Change CreateEventActivity & EventAdapter
Fixed an issue where changing the device's language would not allow events to be filtered by category. The enum name of the Category enum is stored as the category of an event.

The days and hours since a past event occured are now displayed as positive numbers. Added an Math.abs() call in EventAdapter.